backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
playwright==1.40.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncpg==0.29.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0
langchain==0.0.348
openai==1.3.8
google-generativeai==0.3.2
sentry-sdk==1.38.0
prometheus-client==0.19.0
structlog==23.2.0

# Development Dependencies
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
mypy==1.7.1
types-redis==4.6.0.11
types-requests==2.31.0.10
